bump version number for prior change
It will be handy to be able to differentiate between the version of this module that has the "altFilename" handling versus the older version that doesn't (specifically for a module I'm finishing up that relies on this... will be helpful if I can tell people "you must be using at least version 2.2.1 of PageTableExtended"). Thanks!